Extract MarkdownHelper
