+++ app/helpers/application_helper.rb
@@ -1,17 +1,5 @@
-require_dependency 'daimon/render/html'
-
 module ApplicationHelper
-  def render_markdown(markdown_text)
-    markdown = Redcarpet::Markdown.new(Daimon::Render::HTML.new(hard_wrap: true),
-                                       fenced_code_blocks: true,
-                                       tables: true)
-    markdown.render(markdown_text)
-  end
-
-  def extract_plain_text(markdown_text)
-    html = render_markdown(markdown_text.gsub(/#{Page::SEPARATOR}/, ""))
-    Nokogiri::HTML(html).inner_text.gsub(/[[:space:]]+/, " ")
-  end
+  include MarkdownHelper
 
   def google_tag_manager(gtm_id)
     if gtm_id.present?
